requirements.txt
Django==5.1
djangorestframework
requests
numpy
pandas
geopandas
geopy
